#277 finished fields migration
