+++ bb-aem-core/src/main/java/com/cognifide/qa/bb/aem/core/component/dialog/dialogfields/text/JustifyControls.java
@@ -1,6 +1,9 @@
-/*
- * Copyright 2016 Cognifide Ltd..
- *
+/*-
+ * #%L
+ * Bobcat
+ * %%
+ * Copyright (C) 2018 Cognifide Ltd.
+ * %%
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -12,40 +15,20 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ * #L%
  */
 package com.cognifide.qa.bb.aem.core.component.dialog.dialogfields.text;
 
-import static com.cognifide.qa.bb.aem.core.component.dialog.dialogfields.text.ControlToolbarImpl.TOOLBAR_ITEM_SELECTOR;
-
-import com.cognifide.qa.bb.qualifier.PageObject;
 import org.openqa.selenium.WebElement;
-import org.openqa.selenium.support.FindBy;
 
 /**
- * Represents text justification options.
+ * Interface that represents text justification options.
  */
-@PageObject
-public class JustifyControls {
+public interface JustifyControls {
 
-  @FindBy(css = TOOLBAR_ITEM_SELECTOR + "[data-action='justify#justifyleft']")
-  private WebElement justifyLeftBtn;
+  WebElement getJustifyLeftBtn();
 
-  @FindBy(css = TOOLBAR_ITEM_SELECTOR + "[data-action='justify#justifycenter']")
-  private WebElement justifyCenterBtn;
+  WebElement getJustifyCenterBtn();
 
-  @FindBy(css = TOOLBAR_ITEM_SELECTOR + "[data-action='justify#justifyright']")
-  private WebElement justifyRightBtn;
-
-  public WebElement getJustifyLeftBtn() {
-    return justifyLeftBtn;
-  }
-
-  public WebElement getJustifyCenterBtn() {
-    return justifyCenterBtn;
-  }
-
-  public WebElement getJustifyRightBtn() {
-    return justifyRightBtn;
-  }
-
+  WebElement getJustifyRightBtn();
 }
